+++ apps/save-editor/lib/l10n/app_localizations_it.dart
@@ -2106,9 +2106,6 @@ class AppLocalizationsIt extends AppLocalizations {
   String get fallbackItem => 'Oggetto';
 
   @override
-  String get itemNaturalWeapon => 'Arma naturale';
-
-  @override
   String get attributeSkillPointsFallback => 'Punti apprendimento (PA)';
 
   @override