+++ apps/save-editor/lib/l10n/app_localizations_pt.dart
@@ -2102,9 +2102,6 @@ class AppLocalizationsPt extends AppLocalizations {
   String get fallbackItem => 'Item';
 
   @override
-  String get itemNaturalWeapon => 'Arma natural';
-
-  @override
   String get attributeSkillPointsFallback => 'Pontos de aprendizado (PA)';
 
   @override
@@ -5206,9 +5203,6 @@ class AppLocalizationsPtBr extends AppLocalizationsPt {
 
   @override
   String get fallbackItem => 'Item';
-
-  @override
-  String get itemNaturalWeapon => 'Arma natural';
 
   @override
   String get attributeSkillPointsFallback => 'Pontos de aprendizado (PA)';